+++ src/nupic/frameworks/opf/helpers.py
